requirements.txt
fastapi
uvicorn
motor
pymongo
voyageai
openai
python-dotenv
pypdf2
python-multipart
